requirements.txt
requests
beautifulsoup4
lxml
markdownify
python-slugify
PyYAML
